refactor: 공연 접근 권한 커스텀 익셉션에 대한 도메인 책임 변경
