+++ src/main/java/com/prography/lighton/artist/users/presentation/exception/ArtistExceptionHandler.java
@@ -4,7 +4,7 @@ import com.prography.lighton.artist.admin.domain.exception.InvalidApproveStatusT
 import com.prography.lighton.artist.admin.domain.exception.SameApproveStatusException;
 import com.prography.lighton.artist.common.domain.entity.exception.ArtistNotApprovedException;
 import com.prography.lighton.artist.common.domain.entity.exception.ArtistRegistrationNotAllowedException;
-import com.prography.lighton.artist.common.domain.entity.exception.NotAMasterArtistException;
+import com.prography.lighton.performance.common.domain.exception.NotAMasterArtistException;
 import com.prography.lighton.common.utils.ApiUtils;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
@@ -22,12 +22,6 @@ public class ArtistExceptionHandler {
     @ExceptionHandler(ArtistNotApprovedException.class)
     public ResponseEntity<ApiUtils.ApiResult<?>> artistUpdateNotAllowedException(
             ArtistNotApprovedException exception) {
-        return new ResponseEntity<>(exception.body(), exception.status());
-    }
-
-    @ExceptionHandler(NotAMasterArtistException.class)
-    public ResponseEntity<ApiUtils.ApiResult<?>> notAMasterArtistException(
-            NotAMasterArtistException exception) {
         return new ResponseEntity<>(exception.body(), exception.status());
     }
