+++ src/main/java/com/prography/lighton/performance/common/domain/entity/Performance.java
@@ -3,7 +3,7 @@ package com.prography.lighton.performance.common.domain.entity;
 import com.prography.lighton.artist.admin.domain.exception.InvalidApproveStatusTransitionException;
 import com.prography.lighton.artist.admin.domain.exception.SameApproveStatusException;
 import com.prography.lighton.artist.common.domain.entity.Artist;
-import com.prography.lighton.artist.common.domain.entity.exception.NotAMasterArtistException;
+import com.prography.lighton.performance.common.domain.exception.NotAMasterArtistException;
 import com.prography.lighton.common.domain.BaseEntity;
 import com.prography.lighton.genre.domain.entity.Genre;
 import com.prography.lighton.performance.common.domain.entity.association.PerformanceArtist;
