+++ src/main/java/com/prography/lighton/performance/common/domain/exception/NotAMasterArtistException.java
@@ -1,4 +1,4 @@
-package com.prography.lighton.artist.common.domain.entity.exception;
+package com.prography.lighton.performance.common.domain.exception;
 
 import com.prography.lighton.common.utils.ApiUtils;
 import org.springframework.http.HttpStatus;
