+++ src/main/java/com/prography/lighton/performance/users/presentation/exception/PerformanceExceptionHandler.java
@@ -1,7 +1,9 @@
 package com.prography.lighton.performance.users.presentation.exception;
 
+import com.prography.lighton.common.utils.ApiUtils;
 import com.prography.lighton.common.utils.ApiUtils.ApiResult;
 import com.prography.lighton.performance.common.domain.exception.MasterArtistCannotBeRemovedException;
+import com.prography.lighton.performance.common.domain.exception.NotAMasterArtistException;
 import com.prography.lighton.performance.common.domain.exception.PerformanceNotApprovedException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
@@ -19,6 +21,12 @@ public class PerformanceExceptionHandler {
     @ExceptionHandler(MasterArtistCannotBeRemovedException.class)
     public ResponseEntity<ApiResult<?>> masterArtistCannotBeRemovedException(
             MasterArtistCannotBeRemovedException exception) {
+        return new ResponseEntity<>(exception.body(), exception.status());
+    }
+
+    @ExceptionHandler(NotAMasterArtistException.class)
+    public ResponseEntity<ApiUtils.ApiResult<?>> notAMasterArtistException(
+            NotAMasterArtistException exception) {
         return new ResponseEntity<>(exception.body(), exception.status());
     }
 }
